shopglobal_image_search/requirements.txt
python-dotenv==1.0.1
requests==2.31.0
black==23.11.0
flake8==6.1.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
gunicorn==21.2.0
python-multipart==0.0.9
google-cloud-aiplatform==1.93.0
google-cloud-aiplatform[preview]==1.93.0
Pillow==10.2.0